[report] More pythonesc get_highest_ranking.
This new way of retrieving the highest ranking value might be a little bit
slower. But since the RANKING_SCALE might have at maximum 10 keys (worst
scenario imo), it does not really matter.

I thing it is worth to make it way more pythonesc in exchange of a couple of
micro-sec.