[wapiti] Avoid None value when retrieving the vulns.
Ensure that the current Wapiti vuln exists in the `SIGNATURES` dict in order to
avoid None ranking value.

Having a '' default value in the .get() method might generate problema if no
vuln has matched against `SIGNATURES`. Indeed, in that extreme case, it will
returns an empty string instead of a risk id.