[ptp] Breaking rework on PTP parser initialization
Clear separation of PTP initialization and Parser initialization. Now it
is not possible to initialise the parser when instanciating a PTP()
object. This is much easier to maintain since we don't have to hack
around in order to know what parameters are for PTP and what are the
ones for the parser.

The parser is now initialized when calling PTP.parse(args) where args
will be used to instanciate the parser.