PTP now assumes that tools repect semantic versioning.
Adventure time! To make it easy for PTP developers, PTP now relies on
the fact the supported tools are following the semantic versioning
(http://semver.org/).

In other words, tools are not breaking their brackward-compatibility
when only incrementing their MINOR and/or PATCH versions. Assuming so,
PTP should still be able to parse them.

We'll stay vigilent and if a tool does break something, we'll try to fix
the parser (e.g. like for wapiti, see its parse(s)).