[tools] HTTP transactions are now ranked as UNKNOWN
PTP is aiming to provide a unified parsing method for many security
tools in order to easily extract interesting information such as the
ranking of the discovered vulnerabilities.

In addition, PTP has been recently improved to parse the HTTP
transactions generated by different tools. The goal would be then to
link each transaction to the corresponding finding.

In the meantime, it seems more consitant to rank the bulk of
transactions as having a ranking UNKNOWN so that once we assign each
transaction to each finding, there won't be backward-compatibility
break.